app/requirements.txt
asgiref==3.3.4
Django>=2.2
psycopg2>=2.7,<3.0
djangorestframework==3.12.4
PyJWT==2.0.1
pytz==2021.1
sqlparse==0.4.1
